Osascript running as user instead of root
Don’t know what changed in 10.11 but what we found is that the osacript
bits fail when the JSS calls it with a trigger. It seems to fail
because it is trying to auth as root and not the user. This is what we
added and it seems to fix it.